feat(datepicker): implement DatePicker directive
The DatePicker directive (ink-datepicker) is currently in a very
simple state, and does not yet integrate with ng-model, so it is
necessary to make use of callback functions to be notified when the
model changes.

There are currently a few bugs remaining, but I'd like to fix it up
nicely, so that it may be used as declaratively as possible, and
without any issues.